Add files via upload
This project is based on building a Supervised Learning Classification model that will help to identify the potential customers who have a higher probability of churn. Then helps the company to understand the pinpoints and patterns of customer churn and will increase the focus on strategizing customer retention.This project is based on building a Supervised Learning Classification model that will help to identify the potential customers who have a higher probability of churn. Then helps the company to understand the pinpoints and patterns of customer churn and will increase the focus on strategizing customer retention.
Skills: Supervised Learning Classification. (logistic regression · Exploratory Data Analysis and Data Preprocessing · KNN